Increase KCP fragment limit from u8 to u16
- Change frg field from u8 to u16 in packet header (25 bytes total)
- Update encode/decode to use get_u16_le/put_u16_le
- Update Segment struct frg field to u16
- Remove truncating cast in session.rs
- Max message size now ~91MB (65,535 fragments × MTU)
- Internal protocol only, no interop concerns

Nym uses KCP for reliability and multiplexing, not standard real-time
use cases. The u8 limit (255 fragments, ~355KB) was insufficient.

Addresses: nym-yih9

+++ common/nym-kcp/src/session.rs
@@ -50,7 +50,7 @@ pub struct KcpSession {
 #[derive(Debug, Clone)]
 struct Segment {
     sn: u32,
-    frg: u8,
+    frg: u16,
     ts: u32,
     resendts: u32,
     rto: u32,
@@ -60,7 +60,7 @@ struct Segment {
 
 impl Segment {
     #[allow(dead_code)]
-    fn new(sn: u32, frg: u8, data: Vec<u8>) -> Self {
+    fn new(sn: u32, frg: u16, data: Vec<u8>) -> Self {
         Segment {
             sn,
             frg,
@@ -173,7 +173,7 @@ impl KcpSession {
             // first has frg=count-1. This allows receiver to know total count from first packet.
             // In KCP, `frg` is set to the remaining fragments in reverse order.
             // i.e., the last fragment has frg=0, the first has frg=count-1.
-            let frg = (count - i - 1) as u8;
+            let frg = (count - i - 1) as u16;
 
             let seg = Segment {
                 sn: self.snd_nxt,
